+++ src/CoronaDashboard/Services/IChartService.cs
@@ -8,6 +8,6 @@ namespace CoronaDashboard.Services
 {
     public interface IChartService
     {
-        Task<string> HandleRedraw(string label, LineChart<int> chart, Func<IStichtingNice, Task<List<Entry>>> func);
+        Task<string> HandleRedraw(string label, LineChart<double> chart, Func<IStichtingNice, Task<List<Entry>>> func);
     }
 }
